feat: generic db

+++ types/response.go
@@ -8,8 +8,8 @@ type StandardResp[T any] struct {
 type EmptyResp StandardResp[*any]
 
 type ListResp[T any] struct {
-	Count   int `json:"count"`
-	Results []T `json:"results" validate:"dive"`
+	Count   int64 `json:"count"`
+	Results []T   `json:"results" validate:"dive"`
 }
 
 type StandardListResp[T any] StandardResp[ListResp[T]]
